Added script for writing to JSON
This is a simple script that provides the skeleton needed to write information to a JSON file.

Currently can only generate and store one entry, and has a few known bugs. This is less a complete functional program, and more a launchpad for getting the password management component of the project up and functional.