Minor Refactor
Refactored the code a bit to give it far more flexibility than before. Functions for creating files, folders, and loading files.

This will allow the user far more functionality in the future as the project develops. In particular, this will make implementing things like naming password files and folders far, far easier. Plus, the code now is much more readable and clear than it was before.