Added new trim function!
#THINGS DONE
1. Added a trim_passwords function that, as the name suggests, allows the user to trim their passwords. For example, if you don't want to include the "@" symbol, but want the rest of the special characters, you can trim all instances of the character from your password list.

Within the trim_passwords function, when asking for the characters to trim for the user will still giving the option to add more, frustratingly it seems the easiest way to do so was to collect the characters as a list, and the concatenate them into a string. This isn't necessarily ideal, but I couldn't find a better way to do it at the time.

There is almost certainly a more clever way to do this, but I was hard pressed to find out how. Maybe I'll return to it later, but for the time being it works...

(´-εヾ )