Added main.py file.
This is the very humble beginnings of the main.py file that will handle my other files as modules. For now, this is primarily CLI controlled, which isn't the most exciting but its a place to start. I wanted to push this out now so that it was present and showed a demonstration of current progress.

Obviously, more of our functions will be wrapped into our manager/generator as time goes on. 

We're getting there!